Move the post-processing code to its own directory
