+++ ChangeLog
@@ -2,6 +2,7 @@ Version 5.4.0 (XXX 2017)
  * Fix for missing locale info in Windows XP.
  * Empty out the post-processing tables for the any, ady, amy langauges
  * Remove left_print_string() from the API.
+ * Recover pp_lexer.fl from ancient version 2.2!
 
 Version 5.3.16 (15 April 2017)
  * Fix python3 unit tests.
