+++ link-grammar/post-process/pp_knowledge.c
